--- a/Refit.Benchmarks/IPerformanceService.cs
+++ b/Refit.Benchmarks/IPerformanceService.cs
@@ -3,20 +3,20 @@
 public interface IPerformanceService
 {
     [Get("/users")]
-    public Task<string> ConstantRoute();
+    public Task<HttpResponseMessage> ConstantRoute();
 
     [Get("/users/{id}")]
-    public Task<string> DynamicRoute(int id);
+    public Task<HttpResponseMessage> DynamicRoute(int id);
 
     [Get("/users/{id}/{user}/{status}")]
-    public Task<string> ComplexDynamicRoute(int id, string user, string status);
+    public Task<HttpResponseMessage> ComplexDynamicRoute(int id, string user, string status);
 
     [Get("/users/{request.someProperty}")]
-    public Task<string> ObjectRequest(PathBoundObject request);
+    public Task<HttpResponseMessage> ObjectRequest(PathBoundObject request);
 
     [Post("/users/{id}/{request.someProperty}")]
     [Headers("User-Agent: Awesome Octocat App", "X-Emoji: :smile_cat:")]
-    public Task<string> ComplexRequest(int id, PathBoundObject request, [Query(CollectionFormat.Multi)]int[] queries);
+    public Task<HttpResponseMessage> ComplexRequest(int id, PathBoundObject request, [Query(CollectionFormat.Multi)]int[] queries);
 }
 
 public class PathBoundObject