@ChrisPulman 2024-10-26
If ReadAsStringAsync provides the same output long term than existing methods then I see no reason not to use it.